migration-main/requirements.txt
oracledb
apscheduler
openai
python-dotenv
google-generativeai
